requirements.txt
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
wheel==0.44.0
